requirements.txt
# Core scientific stack
numpy>=1.24
pandas>=2.0
scipy>=1.11

# NLP
nltk>=3.8
gensim>=4.3          # NOTE: gensim 4.x removed index2word / syn0 (now index_to_key / vectors)

# Machine learning
scikit-learn>=1.3
joblib>=1.3

# Visualisation
matplotlib>=3.7
wordcloud>=1.9

# Config
pyyaml>=6.0
